extracted the resolving algorithm out of the Import.resolve constructor
Originally, imports were resolved in the Import.resolve factory constructor.
That prevented custom modifications to the algorithm, which might seem like
not a big deal, but one modification actually is very important -- ability
to prevent importing more versions of one package.

So the ImportResolver interface emerges and currently has two implementations:

1. BasicImportResolver, which contains the original resolving algorithm.
   It simply takes whatever is in the repository.

2. RuntimeAwareImportResolver, which knows that at runtime, you really really
   don't want to import two versions of the same package, and prevents this
   by throwing an exception. The hard work is delegated to BasicImportResolver.

The second resolver is now used in resolveLocally.